Relaxes version boundaries, for GHC 7.10 and 8.2.
This commit relaxes the version boundaries of base, directory, process,
and other packages that normally accompany GHC installations, so that
AutoBench works with GHC versions 7.10 to 8.2 (included).

The latest version of Chart available on hackage at this time does not
support GHC 8.4, so this has, for now, been disabled here too.

Closes #3.